fixes smithy-580: improve mapping of finding info
this commit changes the way that fields are mapped to the finding info.
specifically, it sets the finding info Uid to be equal to the rule ID or
rule Guid of the Sarif result. This allows a deduplication engine to use
this id along with the location data to do its job. However, we have
also introduced code to ensure that if each result doesn't have a Guid
based on UUID3 or UUID5, we produce one which is a very simplified type
of deduplication.