fixes saas-623: don't convert relative to absolute paths and validate data source
we are not expecting to parse sarif results with multiple location
entries and we expect all the paths for source code findings to be
paths relative to the root of the repository. This wasn't explicitly
checked until now, we are now enforcing more validity checks and the
filepath.Abs has been replaced with filepath.Clean which will normalise
a relative path without converting it into an absolute one. If a Sarif
result has multiple locations, an error will be returned and if a data
source provided to the transformer will also return an error.