SMITHY-633: fallback to target metadata pURL when finding location parsing fails
there is an edge case where Sarif parsing fails to parse pURLs for OCI
package metadata when these are produced by Trivy. Trivy produces some
findings whose artifact location looks like `workspace/source-code/image.tar`.
When we set the pURL findings ecosystem to `docker` this is converted to
`pkg:docker/workspace/source-code/image.tar` which is an invalid pURL
that causes a panic in our Sarif to OCSF transformer library. We now
handle this error by falling back to the pURL that has been provided to
the library as part of the datasource metadata, meaning the pURL of the
container or artifact that was downloaded. This should provide a more
accurate location for the finding.